requirements.txt
#Core ML & Transformer stack

torch>=1.10.0
transformers>=4.20.0
pytorch-crf>=0.7.2


#Data processing

numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

#NER evaluation

seqeval>=1.2.2
sklearn-crfsuite>=0.3.6

#Utilities

pyyaml>=6.0
tqdm>=4.62.0
python-dotenv>=1.0.0
SentencePiece>=0.1.96

#Portuguese NLP

spacy>=3.4.0    # required for some preprocessing scripts (convert/cleanup)
fasttext         # optional: required by bilstm_fasttext baseline

#Gemini / LLM extraction (optional, only for src/gemini)

google-generativeai
SentencePiece